move findFile from config to general
