+++ applications/compare_cumulative_psf.py
@@ -162,7 +162,7 @@ def main():
     # Measured cumulative PSF
     dataToPlot = OrderedDict()
     if args.data is not None:
-        dataFile = cfg.findFile(args.data)
+        dataFile = gen.findFile(args.data, cfg.get(par="modelFilesLocations"))
         dataToPlot["measured"] = loadData(dataFile)
         radius = dataToPlot["measured"]["Radius [cm]"]
