+++ tests/unit_tests/layout/test_layout_array.py
@@ -6,8 +6,8 @@ import astropy.units as u
 import numpy as np
 import pytest
 
-import simtools.config as cfg
 import simtools.io_handler as io
+import simtools.util.general as gen
 from simtools import db_handler
 from simtools.layout.layout_array import LayoutArray
 
@@ -51,7 +51,7 @@ def telescopeTestFile(db):
         fileName=testFileName,
     )
 
-    cfgFile = cfg.findFile(testFileName, io.getOutputDirectory(dirType="model", test=True))
+    cfgFile = gen.findFile(testFileName, io.getOutputDirectory(dirType="model", test=True))
     return cfgFile
