+++ tests/unit_tests/model/test_mirrors.py
@@ -4,8 +4,8 @@ import logging
 
 import pytest
 
-import simtools.config as cfg
 import simtools.io_handler as io
+import simtools.util.general as gen
 from simtools import db_handler
 from simtools.model.mirrors import Mirrors
 
@@ -27,7 +27,7 @@ def test_read_list(db):
         dest=io.getOutputDirectory(dirType="model", test=True),
         fileName=testFileName,
     )
-    mirrorListFile = cfg.findFile(testFileName, io.getOutputDirectory(dirType="model", test=True))
+    mirrorListFile = gen.findFile(testFileName, io.getOutputDirectory(dirType="model", test=True))
     logger.info("Using mirror list {}".format(mirrorListFile))
     mirrors = Mirrors(mirrorListFile)
     assert 198 == mirrors.numberOfMirrors
